+++ article-skeleton/src/main/java/com/example/article/service/ArticleService.java
@@ -1,7 +1,8 @@
-package com.example.article;
+package com.example.article.service;
 
 import com.example.article.dto.ArticleDto;
 import com.example.article.entity.ArticleEntity;
+import com.example.article.repository.ArticleRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
